Fix a warnings about ignoring vasprintf return code. Factor uses of vasprintf with Stringify::format.
Rationale
---------

On *BSD (including MacOS) `vasprintf` set the buffer pointer to NULL in the event of error
(failed memory allocation). This is not the case with the corresponding GNU extension
(`vasprintf` is a GNU extension of POSIX standard) : the pointer value is undefined when allocation
failed. Instead the return value is set to a negative value (on both *BSD and GNU system).

Current code only checks for NULL pointer values. GCC warns about the ignored return code.

There is multiple uses of vasprint in the code :
* `Stringigy::format()` is the abstraction that should be used everywhere for wrapping the `vasprintf` functionality for c++ code,
* `StringLine::format()`,
* `CollectionAbstract::addProperty()`,
* `Graph::create()` and the duplicated `GraphUnitig::create()`.

All locations trigger the warning (as of GCC 6.3.1 on 4.9.8-1-ARCH x86_64 GNU/Linux).

GATB is run on systems with lots of memory pressure. This error may forbid programs to fail nicely in extremes cases.
It is *not* critical.

Course of action
----------------

Return code is checked to be > -1. Also, the check for NULL pointer value is conservatively kept.

Instead of repeating the change multiple times, every uses of `vasprintf` are replaced with a call to
`Stringigy::format()`. An new overloaded and non-variadic `Stringigy::format()` accepts a `va_list` parameter,
allowing variadic functions to delegate to it. This is similar to `*printf` and like functions that allow both
variadic and non-variadic arguments.

Result
------

No warnings. Only one use of `vasprintf`.