Fix a rounding error in Data assignment when size() is a multiple of 4
The number of bytes read when a BINARY Data instance is copied was (NbNucletotides/4)+1.
The correct value is (NbNucletotides+3)/4. When NbNucletotides % 4 == 0, one
byte is read outside the buffer.

Make clear in the doc that the docummentation of the inherited Vector::size() lies.

Added a getBufferLength method and changed the faulty line for using it.